fix(request): Fixed request end method .toLowerCase() error
When request is real (request is using original superagent methods) and having no headers set, the `Request.prototype.end()` method was broken. The original `Request.prototype.set` method was mocked and prevented the tests from failing and detecting this case. So the tests were updated, we no longer mock the original `set` method and a fallback in case `request.headers` is undefined is provided.